client/requirements.txt
pyepics>=3.5.0
numpy>=1.24.0
matplotlib>=3.7.0
